Fix: Blender crashes with emissive triangles
This is related to the light-tree implementation.
The issue most often happens when the emissive material is just added
and the preview material is then rendered.

The crash is caused by iterating through the wrong thing.
We are now iterating through emissive primitives, and
many of these primitives correspond to the same object.
However, the inner loop was still iterating through all triangles of the
object, so it was iterating through the same object multiple times.