RenderReport: Add option to add platform specific overrides.
Reference images in the platform_reference_dir will be chosen before
images in reference_dir. This allows platform specific reference
images, with a common base.

To ignore set platform_reference_dir to None. The caller is responsible
of setting the platform reference dir as the unit test is more aware
what the definition of a platform is.

NOTE: should we rename it to reference_override_dir as that is better
suited to the actual implementation and responsibilities.

Differential Revision: https://developer.blender.org/D15265