Fix T93650: Asset drag into catalogs broken, Industry Compatible keymap
Issue was that the Industry Compatible keymap wouldn't select files on a
mouse press, and since the dragged items are determined by the
selection, nothing would be considered as dragged.

Selecting items on mouse press happens since c606044157a3. I haven't
heard of that issue happening in the Industry Compatible keymap. But if
it did happen, it should be fixed too now.