Fix: Cycles light tree triangle sampling
This currently reverts some of the distant light sampling changes,
so that I could isolate the issues with triangle light sampling.

The fix ensures that the light distribution still computes the correct
cdf (if needed), and removes the light tree kernel's dependency on the
light distribution. This also corrects the triangle PDFs.