Curve sculpting: Fixes and optimizations for collision solver.
- Use the actual curve radius instead of a hardcoded value.
  This currently has to be set in a geometry node tree and then applied.
- Optimization: ignore collisions for anchored/pinned points.
- Fixed jittery collision solve and exploit point order for performance.

The collision solution was placing points at 2x the required radius
from the surface and then did not detect new collision until in range
of 1 radius from the surface. This caused jaggy movement when pulling
points toward the surface. The root cause was that the raycast callback
already includes the radius as an offset of the hit point, so adding it
in the Laplace multiplier effectively applies the margin twice. Leaving
it out of lambda calculation fixes the issue.

Constraint solving can be optimized further by solving for each point
along the curve without violating constraints of the previous point.
This "Follow the Leader" approach works because of the anchored root
point.