Curve sculpting: Fix threading issues in constraint solver.
The constraint solver gets a span of changed curve indices. The comb
brush collects these in a thread-local vector, which is only valid in
a `parallel_for` loop.

The BVH storage also needs to be local to the function, otherwise
parallel calls to the constraint solver will free the shared storage
at the end of the task.