Curve sculpt: Improvements and testing for collision constraint solver.
Collision is now optional in the curve sculpting tools and disabled by
default (button next to X/Y/Z symmetry settings).

Moved constraint solver into blenkernel for wider accessibility.

Changed collision method from raycasts to substeps with sphere casts.
This method of finding and resolving contacts is recommended by the XPBD
author. It avoids continuous collision detection (CCD) with raycasts
in favor of using a number of substeps. The range of potential contacts
is limited by the size of the overall allowed step size and number of
substeps.

The length constraints now work in both directions except for the first
segment, which avoids numerical stiffness close to the root.

Constraints now follow a stricter general recipe, which should make it
easier to implement future constraints with more complex constraint
functions and gradients.

The solver now outputs results with overall numerical error metrics and
timings.

Added a basic performance test for the curves constraint solver. The
test is parameterized for over a range of substep counts.