Animation: Add redo panel to Dopesheet and NLA
Animation: Add redo panel to Dopesheet and NLA

---

Also implicitly adds it for the timeline editor, since it's a kind of action editor internally.

This feature is needed for changing advanced properties of animation operators, such as select grouped (see D14811).
But it can also be useful for existing operators, like precise keyframe position tweaking.

Changes are basically the same as in D6286 (which added the redo panel for Graph Editor).

Some operators have internal properties that should be hidden. A full list can be found in T98195. These will be fixed in a follow-up patch.

{F13079611} {F13079612}

Reviewed By: sybren

Differential Revision: https://developer.blender.org/D14960