Animation: Fix operator properties for redo panel
Animation: Fix operator properties for redo panel

After the redo panel is added to animation editors in D14960, many operators
must be adjusted to appear and function correctly.

A full list of changes is tracked in T98195

---

This patch only includes actual usability fixes. It does not do any changes for the user's convenience, like adding other helpful properties to operators. This can be done in a follow-up patch.

Reviewed By: sybren

Maniphest Tasks: T98195

Differential Revision: https://developer.blender.org/D14977