Blender build system changes for new shared libraries and modules on Linux/macOS
PLATFORM_BUNDLED_LIBRARY_DIRS is added to the rpath for the build steps so
no libraries have to be copied.

PLATFORM_BUNDLED_LIBRARIES are copied to the lib folder in the install step.
The Blender gets its rpath point to this lib older as part of the install
step. The Python executable already contains an rpath to the lib folder.

For site packages installation on Linux, just copy over the entire library
directory as we do for macOS to simplify things.