Stop statically linking libstdc++ on Linux
With the addition of shared libraries this doesn't work anymore. The shared
libraries use dynamical linking, and having each library with their own static
copy seems problematic.

The use of this predates our switch to CentOS 7. I hope that there is just
no need for this anymore, but it's not entirely clear.