Fix an illegal address and assert issue with Many Lights Sampling
If the emitter_index in light_tree_emitter_reservoir_weight() is less than 0,
then this lead to an error when using the kernel_data_fetch function.

This patch fixes this by returning a weight of 0 and ending the function before
we reach kernel_data_fetch in this situation.  Returning a weight of 0 is safe
here as an emitter_index less than 0 indicates that no lights were picked. So
they will have a weight of 0.

Contributed by Alaska

Differential Revision: https://developer.blender.org/D15991