Fix distant lights energy calculation
We currently have an issue where distance lights are infrequently
sampled when using Many Lights Sampling.
One cause for this is that distance lights are assigned relatively
low importance compared to other light types with the same
lighting contribution to a given object/material.

This appears to be because we are using the "strength" of the distant
light as our energy when in fact we should be using the energy, which
is roughly `4pi * strength`.

Differential Revision: https://developer.blender.org/D16148