Fix PDF calculation when splitting is enabled
Previously the PDF calculation while splitting and considering MIS was incorrect.
If we reached a node that did not meet the criteria to split,  then the PDF calculation would resort to following the bit trail to the target leaf as if the current node was the root node. If the light tree split at least once, then this approach was almost always going to be incorrect.
Along with this, we also want to traverse through the light tree until we reach a random leaf under certain circumstances. The previous code did not have anything in place for this which also lead to some incorrect results.

In this patch, I attempted to fix both of these issues by doing these things when we encounter a node that does not meet the criteria to split:
1. Follow the bit trail starting from the position of the current node, rather than following the bit trail from the root node when we're in fact starting from a different node.
2. Guarantee that the current node is along the path of the bit trail before following the path of the bit trail.
3. And if the current node is not along the path, then we traverse down the tree to a random leaf node.

Differential Revision: https://developer.blender.org/D15983