deps_builder: Silence CMP0135 warnings
CMake 3.24+ has the option to set the time
stamp of extracted file to the time of
extraction rather than the time in the
archive. Which is nice, it also however
added a multi line warning when you do
not explicitly set this option. So that
got a little noisy.

This change explicitly picks the new
policy and thus silences the warning.