deps_builder: limit llvm/dpcpp to c++14
We build all deps with the -DCMAKE_CXX_STANDARD=17
flag. llvm however will not switch over to C++17
until llvm 16.0, and building older versions with
this flag is leading to some crashes inside ispc
when building with MSVC

So for both llvm and dpcpp remove the -DCMAKE_CXX_STANDARD=17
flag, this is done for all platforms since we should
not be overriding llvm's preferences here.