Deps builder: Initial pass as deps update for 3.4
This was mostly just bumping deps with known CVE's
the only exception was OIIO/OSL. OIIO needed a rebuild
anyhow because of openjpeg, so was it was no effort to
bump it to 2.4.x where some of the work @aras_p has
been doing had been landed, also I think there were some
drastic perf improvements with one of the USD sample scenes.

|dep|old|new|
|zlib|1.2.12|1.2.13|
|freetype|2.11.1|2.12.1|
|openimageio|2.3.13.0|2.4.4.2|
|OSL|1.11.17.0|1.12.6.2|
|python|3.10.2|3.10.8|
|openjpeg|2.4.0|2.5.0|
|ffmpeg|5.0|5.1.2|
|sndfile|1.0.28|1.1.0|
|xml2|2.9.10|2.10.3|
|expat|2.4.4|2.4.9|
|openssl|1.1.1g/i|1.1.1q|
|sqlite|3.31.1|3.37.2|
|openpgl|0.3.1-beta|0.4.0-beta|

This diff is mostly just so we can get an easy overview of the
changes being done, please do any work in the `tmp_libs_34` branch.
(and update this diff occasionally)

Notable changes:

AOM : the hack we had in place to make it not detect
pthreads on windows no longer worked with a more
recent cmake version. Disabled pthreads with a diff
(windows only)

Python:
Python had an embedded copy of zlib 2.1.12 swapped it
out for our 2.1.13 copy with some folder manipulation
on windows.

Freetype:
Freetype was harbouring a copy of zlib 2.1.12 as well,
so that had to end.

FFMpeg:
The patch we had in place for D11796 no longer applies,
and it doesn't look like it is needed any-more, but i
was unable to verify this since the original problem
only showed on mac.

There is a new patch to deal with simple_idct.asm
generating an object file with no sections in it,
upsetting strip, causing a build error. This is a
backport from an upstream commit.

SQLITE:
They changed their filenames a bit, given python takes
care of these on windows, i hope i picked the right one.

OSL:
OSL has an extra lib now: oslnoise. this may need
adjustments in the platform/findXXXX cmake files for that.

TINYXML:
was still being downloaded, but nothing seemingly uses
it, as there are no build scripts for it. Removed it
from versions/download.cmake