More accurate light behind surface test using all corners of bounding box
In Many Lights Sampling we use whether or not a bounding box is
behind an opaque surface to decide whether or not we discard or
consider the bounding box.

However the calculation used for figuring out if a bounding box
is behind a surface ended up being wrong in some situations.

This patch attempts to fix this by checking if each corner of
the bounding box is behind the surface, resulting in a (hopefully)
perfect differentiation between boxes behind and infront of a
surface.

Contributed by Alaska, with changes by Brecht.

Differential Revision: https://developer.blender.org/D16192