Store ID in asset representation directly
This should make some things more convenient in future, since we can
directly access the ID if an asset represents one, without having to
look it up. But also, we need to do ID remapping either way, which
currently requires the ID pointer to identify the ID, not other ID data
(like the contained asset metadata).