BLI: Refactor Matrix C++ API to be fully templated
This new API tries to make everything templated so it can be used in more places.

The scope of this patch is to transfer all the functionality currently exposed by `float4x4` and  `float3x3` without relying on any BLI C function.
This was a good opportunity to check the amount of work needed for a deep rewrite and create a solid base to build upon.

This adds new Rotation types in order to clean-up the API. This will be extended when we port the full Rotation API. The types are made so that they don't allow implicit
down-casting to their vector representation.

Some functions make direct use of the Eigen library, bypassing the Eigen C API defined in `intern/eigen`. Its use is contained inside `math_matrix.cc`. There is
conflicting opinion wether we should use it more so I contained its usage to almost the tasks as in the C API for now.

I could not find a way to have the `operator*` for matrix multiplication inside the `mat_base` class and be able to define the template specialization for SSE2.

Todo:
- Matrix view for some operations. This is mostly needed when we need to work on only the 3x3 portion of a 4x4 matrix, or when working with bigger matrices. I am unsure
what would be the best way to achieve this. It might be better to leave it to another patch.

Some consideration:
- Should the single scalar constructor (i.e: `float3x3(1.0f)`) be allowed and what should it do? Currently it matches the GLSL constructor which initialize the diagonal of
the matrix with this value.
- The static init helpers are hidden in a specific namespace (i.e: `blender::math::mat3x3`). This is because it allows to use this namespace for more concise code. But
`mat3x3` could also be put as prefix or sufix in the function name.

Differential Revision: https://developer.blender.org/D16625