Add initial version of Node Sections to keep sockets more manageable
Nodes can now have sections, and each socket can belong to a section.

These only affect how the node is displayed. In the Node Editor, the node first
shows all sockets without a section as usual, and following that the sections
are shown.
Each section can be expanded or collapsed, hiding the sockets inside.

Node links are still drawn if the socket they're going to is invisible -
in that case, they'll be going to the section header instead. However,
dragging them or connecting a link to a hidden socket via dragging is not
possible.

Node Groups can also have sections defined by the user, which are reflected
in their group input/output and in the group nodes that use them.

Note that the code for hidden nodes is still missing, I'm not sure yet how
to handle this there.
Also, we might eventually support drawing node properties (such as SSS type
for the principled BSDF) inside a section.
Similarly, especially for custom groups it might be nice to support nested
sections.